Force Briar to be last in All Bosses
This doesn't have much of an effect on logic, and I'm doing it so that you can end your run with the actual final boss. It'd be good to add a gate in front of Briar that requires you to have beaten all of the other bosses first (only if the goal is All Bosses, though).

The only real effect this has on logic is that vanilla postgame mode forces all postgame content behind the win condition.